backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
ccxt==4.4.54
pandas==2.2.3
pandas-ta==0.3.14b1
sqlalchemy==2.0.36
aiosqlite==0.20.0
websockets==14.1
python-dotenv==1.0.1
httpx==0.28.1
pydantic==2.10.4
